Rename c_bindings to ffi.
